Create a 1.24-series branch

+++ _travis/upload_coverage.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
-set -e
-set -x
+set -exo pipefail
 
-source .tox/${TOXENV}/bin/activate
-pip install codecov
-codecov --env TRAVIS_OS_NAME,TOXENV
+if [[ -e .coverage ]]; then
+    source .tox/${TOXENV}/bin/activate
+    pip install codecov
+    codecov --env TRAVIS_OS_NAME,TOXENV
+fi
